PHAD-Brain/requirements.txt
# Core
numpy>=1.26,<2.1
scipy>=1.11,<1.13
typing_extensions>=4.8
dataclasses-json>=0.6

# Messaging
pyzmq>=25.1

# FRC / NetworkTables (RobotPy)
robotpy-ntcore>=2024.3.1
robotpy-wpilib-utilities>=2024.3.1

# Vision
opencv-python>=4.9
pillow>=10.2

# ML (CPU)
torch>=2.2
torchvision>=0.17
tensorflow-cpu>=2.15

# Model export
onnx>=1.16
onnxruntime>=1.17

# Utils
tqdm>=4.66
pyyaml>=6.0
rich>=13.7
